[MOVE] commands -> commands.legacy
