+++ src/main/java/com/github/unchama/seichiassist/commands/legacy/MebiusCommand.java
@@ -1,4 +1,4 @@
-package com.github.unchama.seichiassist.commands;
+package com.github.unchama.seichiassist.commands.legacy;
 
 import com.github.unchama.seichiassist.listener.MebiusListener;
 import org.bukkit.ChatColor;
